Unnecessary unboxing
Reports "unboxing", e.g. explicit unwrapping of wrapped primitive values. Unboxing is unnecessary under Java 5 and newer, and can be safely removed.

+++ asm/src/main/java/org/aspectj/asm/AsmManager.java
@@ -1191,7 +1191,7 @@ public class AsmManager implements IStructureModel {
 			if (ctr == null) {
 				nodeTypeCount.put(node, 1);
 			} else {
-				ctr = ctr.intValue() + 1;
+				ctr = ctr + 1;
 				nodeTypeCount.put(node, ctr);
 			}
 		}
