+++ asm/src/main/java/org/aspectj/asm/internal/JDTLikeHandleProvider.java
@@ -172,7 +172,7 @@ public class JDTLikeHandleProvider implements IElementHandleProvider {
 						String existingHandle = object.getHandleIdentifier();
 						int suffixPosition = existingHandle.indexOf('!');
 						if (suffixPosition != -1) {
-							count = new Integer(existingHandle.substring(suffixPosition + 1)).intValue() + 1;
+							count = new Integer(existingHandle.substring(suffixPosition + 1)) + 1;
 						} else {
 							if (count == 1) {
 								count = 2;
@@ -228,7 +228,7 @@ public class JDTLikeHandleProvider implements IElementHandleProvider {
 							String existingHandle = object.getHandleIdentifier();
 							int suffixPosition = existingHandle.indexOf('!');
 							if (suffixPosition != -1) {
-								count = new Integer(existingHandle.substring(suffixPosition + 1)).intValue() + 1;
+								count = new Integer(existingHandle.substring(suffixPosition + 1)) + 1;
 							} else {
 								if (count == 1) {
 									count = 2;
@@ -280,7 +280,7 @@ public class JDTLikeHandleProvider implements IElementHandleProvider {
 							String existingHandle = object.getHandleIdentifier();
 							int suffixPosition = existingHandle.indexOf('!');
 							if (suffixPosition != -1) {
-								count = new Integer(existingHandle.substring(suffixPosition + 1)).intValue() + 1;
+								count = new Integer(existingHandle.substring(suffixPosition + 1)) + 1;
 							} else {
 								if (count == 1) {
 									count = 2;
@@ -315,7 +315,7 @@ public class JDTLikeHandleProvider implements IElementHandleProvider {
 							int suffixPosition = existingHandle.lastIndexOf('!');
 							int lastSquareBracket = existingHandle.lastIndexOf('['); // type delimiter
 							if (suffixPosition != -1 && lastSquareBracket < suffixPosition) { // pr260384
-								count = new Integer(existingHandle.substring(suffixPosition + 1)).intValue() + 1;
+								count = new Integer(existingHandle.substring(suffixPosition + 1)) + 1;
 							} else {
 								if (count == 1) {
 									count = 2;
@@ -335,7 +335,7 @@ public class JDTLikeHandleProvider implements IElementHandleProvider {
 							int suffixPosition = existingHandle.lastIndexOf('!');
 							int lastSquareBracket = existingHandle.lastIndexOf('['); // type delimiter
 							if (suffixPosition != -1 && lastSquareBracket < suffixPosition) { // pr260384
-								count = new Integer(existingHandle.substring(suffixPosition + 1)).intValue() + 1;
+								count = new Integer(existingHandle.substring(suffixPosition + 1)) + 1;
 							} else {
 								if (count == 1) {
 									count = 2;
@@ -384,7 +384,7 @@ public class JDTLikeHandleProvider implements IElementHandleProvider {
 					String existingHandle = object.getHandleIdentifier();
 					int suffixPosition = existingHandle.indexOf('!');
 					if (suffixPosition != -1) {
-						count = new Integer(existingHandle.substring(suffixPosition + 1)).intValue() + 1;
+						count = new Integer(existingHandle.substring(suffixPosition + 1)) + 1;
 					} else {
 						if (count == 1) {
 							count = 2;
