+++ testing/src/test/java/org/aspectj/testing/xml/SoftMessage.java
@@ -294,7 +294,7 @@ public class SoftMessage implements IMessage {
 		if (null != sourceLocation) {
 			throw new IllegalStateException("cannot set line after creating source location");
 		}
-		this.line = Integer.valueOf(line).intValue();
+		this.line = Integer.valueOf(line);
 		SourceLocation.validLine(this.line);
 	}
